Fix linux process helper and properly adjust macos process helper

+++ daemon/platform/linux/process_helper.rs
@@ -20,6 +20,14 @@ pub fn compile_shell_command(command_string: &str) -> Command {
 
 /// Send a signal to one of Pueue's child process handles.
 /// We need a special since there exists some inconsistent behavior.
+///
+/// In some circumstances and environments `sh -c $command` doesn't spawn a shell,
+/// but rather spawns the `$command` directly.
+///
+/// This makes things a lot more complicated, since we need to either send signals
+/// to the root process directly OR to all it's child processes.
+/// This also affects the `--children` flag on all commands. We then have to either send the signal
+/// to all direct children or to all of the childrens' children.
 pub fn send_signal_to_child(
     child: &Child,
     action: &ProcessAction,
@@ -27,18 +35,31 @@ pub fn send_signal_to_child(
 ) -> Result<bool> {
     let signal = get_signal_from_action(action);
     let pid: i32 = child.id().try_into().unwrap();
+    // Check whether this process actually spawned a shell.
+    let is_shell = if let Ok(is_shell) = did_process_spawn_shell(pid) {
+        is_shell
+    } else {
+        return Ok(false);
+    };
 
-    // There might be multiple children, for instance, when users use the `&` operator.
-    // If the `send_to_children` flag is given, the
+    if is_shell {
+        // If it's a shell, we have to send the signal to the actual shell and to all it's children.
+        // There might be multiple children, for instance, when users use the `&` operator.
+        // If the `send_to_children` flag is given, the
 
-    // Send the signal to the shell, don't propagate to it's children yet.
-    send_signal_to_process(pid, action, false)?;
+        // Send the signal to the shell, don't propagate to it's children yet.
+        send_signal_to_process(pid, action, false)?;
 
-    // Now send the signal to the shells child processes and their respective
-    // children if the user wants to do so.
-    let shell_children = get_child_processes(pid);
-    for shell_child in shell_children {
-        send_signal_to_process(shell_child.pid(), action, send_to_children)?;
+        // Now send the signal to the shells child processes and their respective
+        // children if the user wants to do so.
+        let shell_children = get_child_processes(pid);
+        for shell_child in shell_children {
+            send_signal_to_process(shell_child.pid(), action, send_to_children)?;
+        }
+    } else {
+        // If it isn't a shell, send the signal directly to the process.
+        // Handle children normally.
+        send_signal_to_process(pid, action, send_to_children)?;
     }
 
     signal::kill(Pid::from_raw(pid), signal)?;
@@ -53,10 +74,20 @@ pub fn send_signal_to_child(
 pub fn kill_child(task_id: usize, child: &mut Child, kill_children: bool) -> bool {
     let pid: i32 = child.id().try_into().unwrap();
 
+    // Check whether this process actually spawned a shell.
+    let is_shell = if let Ok(is_shell) = did_process_spawn_shell(pid) {
+        is_shell
+    } else {
+        return false;
+    };
+
     // We have to kill the root process first, to prevent it from spawning new processes.
     // However, this prevents us from getting it's child processes afterwards.
     // That's why we have to get the list of child processes already now.
-    let child_processes = get_child_processes(pid);
+    let mut child_processes = None;
+    if kill_children || is_shell {
+        child_processes = Some(get_child_processes(pid));
+    }
 
     // Kill the parent first
     match child.kill() {
@@ -66,22 +97,69 @@ pub fn kill_child(task_id: usize, child: &mut Child, kill_children: bool) -> boo
         }
         _ => {
             // Now kill all remaining children, after the parent has been killed.
-            for child_process in child_processes {
-                // Send the signal to each child process, show warning if this fails.
-                let process_pid = child_process.pid();
-                if let Err(error) =
-                    send_signal_to_process(process_pid, &ProcessAction::Kill, kill_children)
-                {
-                    warn!(
-                        "Failed to send kill to pid {} with error {:?}",
-                        process_pid, error
-                    );
+            // If a shell is spawned, we have to manually send the kill signal to all children.
+            // Otherwise only send a signal to all children if the `kill_children` flag is set.
+            if let Some(child_processes) = child_processes {
+                if is_shell {
+                    for child_process in child_processes {
+                        // Send the signal to each child process, show warning if this fails.
+                        let process_pid = child_process.pid();
+                        if let Err(error) =
+                            send_signal_to_process(process_pid, &ProcessAction::Kill, kill_children)
+                        {
+                            warn!(
+                                "Failed to send kill to pid {} with error {:?}",
+                                process_pid, error
+                            );
+                        }
+                    }
+                } else if kill_children {
+                    send_signal_to_processes(child_processes, &ProcessAction::Kill);
                 }
             }
 
             true
         }
     }
+}
+
+/// Check whether a process's commandline string is actually a shell or not
+fn did_process_spawn_shell(pid: i32) -> Result<bool> {
+    // Get the /proc representation of the child, so we can do some checks
+    let process = if let Ok(process) = Process::new(pid) {
+        process
+    } else {
+        info!(
+            "Process to kill has probably just gone away. Process {}",
+            pid
+        );
+        bail!("Process has just gone away");
+    };
+
+    // Get the root command and, so we check whether it's actually a shell with `sh -c`.
+    let mut cmdline = if let Ok(cmdline) = process.cmdline() {
+        cmdline
+    } else {
+        info!(
+            "Process to kill has probably just gone away. Process {}",
+            pid
+        );
+        bail!("Process has just gone away");
+    };
+
+    if cmdline.len() < 3 {
+        return Ok(false);
+    }
+
+    if cmdline.remove(0) != "sh" {
+        return Ok(false);
+    }
+
+    if cmdline.remove(0) != "-c" {
+        return Ok(false);
+    }
+
+    Ok(true)
 }
 
 /// Send a signal to a unix process.
@@ -185,6 +263,9 @@ mod tests {
         // Sleep a little to give everything a chance to spawn.
         sleep(Duration::from_millis(500));
 
+        // Make sure the process indeed spawned a shell.
+        assert!(did_process_spawn_shell(pid).unwrap());
+
         // Get all child processes, so we can make sure they no longer exist afterwards.
         let child_processes = get_child_processes(pid);
         assert_eq!(child_processes.len(), 2);
@@ -214,6 +295,9 @@ mod tests {
         let pid: i32 = child.id().try_into().unwrap();
         // Sleep a little to give everything a chance to spawn.
         sleep(Duration::from_millis(500));
+
+        // Make sure the process indeed spawned a shell.
+        assert!(did_process_spawn_shell(pid).unwrap());
 
         // Get all child processes and all childrens children,
         // so we can make sure they no longer exist afterwards.
@@ -256,6 +340,9 @@ mod tests {
         // Sleep a little to give everything a chance to spawn.
         sleep(Duration::from_millis(500));
 
+        // Make sure the process did not spawn a shell.
+        assert!(!did_process_spawn_shell(pid).unwrap());
+
         // No childprocesses exist
         let child_processes = get_child_processes(pid);
         assert_eq!(child_processes.len(), 0);
@@ -281,6 +368,9 @@ mod tests {
         let pid: i32 = child.id().try_into().unwrap();
         // Sleep a little to give everything a chance to spawn.
         sleep(Duration::from_millis(500));
+
+        // Make sure the process indeed spawned a shell.
+        assert!(!did_process_spawn_shell(pid).unwrap());
 
         // Get all child processes, so we can make sure they no longer exist afterwards.
         let child_processes = get_child_processes(pid);
